+++ src/Products/PloneMeeting/Meeting.py
@@ -17,6 +17,7 @@ from collections import OrderedDict
 from collective.behavior.talcondition.utils import _evaluateExpression
 from collective.contact.plonegroup.config import get_registry_organizations
 from collective.eeafaceted.dashboard.utils import enableFacetedDashboardFor
+from collective.iconifiedcategory.interfaces import IIconifiedPreview
 from copy import deepcopy
 from DateTime import DateTime
 from DateTime.DateTime import _findLocalTimeZoneName
@@ -47,6 +48,7 @@ from Products.CMFCore.permissions import ReviewPortalContent
 from Products.CMFCore.utils import _checkPermission
 from Products.CMFDynamicViewFTI.browserdefault import BrowserDefaultMixin
 from Products.CMFPlone.utils import base_hasattr
+from Products.CPUtils.Extensions.utils import remove_generated_previews
 from Products.PageTemplates.Expressions import SecureModuleImporter
 from Products.PloneMeeting.browser.itemchangeorder import _compute_value_to_add
 from Products.PloneMeeting.browser.itemchangeorder import _is_integer
@@ -64,6 +66,8 @@ from Products.PloneMeeting.utils import display_as_html
 from Products.PloneMeeting.utils import displaying_available_items
 from Products.PloneMeeting.utils import fieldIsEmpty
 from Products.PloneMeeting.utils import forceHTMLContentTypeForEmptyRichFields
+from Products.PloneMeeting.utils import fplog
+from Products.PloneMeeting.utils import get_annexes
 from Products.PloneMeeting.utils import getCustomAdapter
 from Products.PloneMeeting.utils import getDateFromDelta
 from Products.PloneMeeting.utils import getFieldVersion
@@ -270,6 +274,25 @@ class MeetingWorkflowActions(object):
                                                   name='pm_unrestricted_methods')
         self.context.setFirstItemNumber(unrestrictedMethodsView.findFirstItemNumberForMeeting(self.context))
         self.context.updateItemReferences()
+        # remove annex previews of every items if relevant
+        if self.cfg.getRemoveAnnexesPreviewsOnMeetingClosure():
+            # add logging message to fingerpointing log
+            for item in self.context.getItems(ordered=True):
+                annexes = get_annexes(item)
+                if annexes:
+                    for annex in annexes:
+                        remove_generated_previews(annex)
+                        annex_infos = item.categorized_elements.get(annex.UID())
+                        if annex_infos:
+                            annex_infos['preview_status'] = IIconifiedPreview(annex).status
+                    item._p_changed = True
+                extras = 'item={0} number_of_annexes={1}'.format(repr(item), len(annexes))
+                fplog('remove_annex_previews', extras=extras)
+            msg = translate(
+                u"Preview of annexes were deleted upon meeting closure.",
+                domain='PloneMeeting',
+                context=self.context.REQUEST)
+            api.portal.show_message(msg, request=item.REQUEST)
 
     security.declarePrivate('doPublish_decisions')