+++ src/Products/PloneMeeting/profiles/__init__.py
@@ -489,6 +489,7 @@ class MeetingConfigDescriptor(Descriptor):
         self.itemCreatedOnlyUsingTemplate = False
         self.annexToPrintMode = 'enabled_for_info'
         self.keepOriginalToPrintOfClonedItems = True
+        self.removeAnnexesPreviewsOnMeetingClosure = False
         self.cssClassesToHide = 'highlight\nhighlight-red'
         self.hideCssClassesTo = ()
         self.enableItemDuplication = True